refactor: abstracts away repositories
