+++ core-commonui/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/core/commonui/modals/SortBottomSheet.kt
@@ -48,8 +48,6 @@ class SortBottomSheet(
         SortType.NewComments,
         SortType.MostComments,
         SortType.Old,
-        SortType.Controversial,
-        SortType.Scaled,
         SortType.Top.Generic,
     ),
     private val expandTop: Boolean = false,
